Add top level Manager project to Monitor
