+++ Source/Client/CLI/Recipe.toml
@@ -20,7 +20,7 @@ Runtime = [
 	"../../Tools/Mkdir/",
 	# "../../Extensions/Cpp/Extension/",
 	# "../../Extensions/CSharp/Extension/",
-	"../../Monitor/Shared/",
+	"../../Monitor/Manager/",
 ]
 Build = [
 	# TODO: "../../Extensions/TestBuild/",
