+++ Source/Monitor/Manager/DetourCallbackLogger.h
