+++ Source/Monitor/Manager/ScopedDetourProcessManagerRegister.h
